requirements.txt
Flask==2.0.2
PyJWT==2.3.0
Werkzeug==2.0.2
jsonschema==4.4.0
SQLAlchemy==1.4.31
alembic==1.7.5